+++ db/revert/backfill_archived_statuses.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:backfill_archived_statuses from pg
+
+BEGIN;
+
+-- XXX Add DDLs here.
+
+COMMIT;