+++ db/revert/computed_columns/application_analyst_status.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:computed_columns/application_analyst_status from pg
+
+begin;
+
+drop function ccbc_public.application_analyst_status;
+
+commit;