+++ db/revert/import/application_status_archived.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:import/set_application_status_archived from pg
+
+BEGIN;
+
+drop function ccbc_public.set_application_status_archived;
+
+COMMIT;