+++ db/revert/tables/application_gis_data.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:tables/application_gis_data from pg
+
+BEGIN;
+
+    drop table ccbc_public.application_gis_data;
+
+COMMIT;