+++ db/revert/tables/application_status_type_005_add_applicant_conditionally_approved.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:tables/application_status_type_005_add_applicant_conditionally_approved from pg
+
+BEGIN;
+
+-- XXX Add DDLs here.
+
+COMMIT;