+++ db/revert/util_functions/upsert_archive_trigger.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:util_functions/upsert_archive_trigger from pg
+
+BEGIN;
+
+drop function if exists upsert_archive_trigger(text, text);
+
+COMMIT;